pyproject.toml
[tool.poetry]
name = "anchor-collateral-steth"
version = "0.1.0"
description = "Repository for contains Ethereum contracts for integrating stETH as a collateral into the Anchor protocol."
authors = ["Lido devteam"]
license = "MIT"
readme = "README.md"
homepage = "https://mainnet.lido.fi"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
eth-brownie = "~1.19.3"
setuptools = "~65.5.1"

[tool.poetry.dev-dependencies]
black = "~22.10.0"
pre-commit = "~2.19.0"
pytest-check = "^2.1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = [
    "tests",
]